Switch hombrew formula to use full archive
I did this because before it was just using the binary and this was
preventing me from including the man page in future versions. Also,
doing the binary the way I was isn't the officially supported binary
mechanism. They intend for bottles to be used to manage binaries which I
plan on doing in the future.