Release git-ps-rs v3.3.0 and remove binary bottles
Release git-ps-rs v3.3.0 and remove the binary bottles because I don't
think they are providing that much value since we have to have gpgme as
a runtime dependency because we haven't figured out how to statically
link gpgme as a dependency. If we figure that out then I will likely
bring the binary bottles back. Also if more of our users have to have
rust installed on their system they are closer to being able to
contribute.